pyproject.toml
[project]
name = "open-strix"
version = "0.1.19"
description = "Minimal autonomous agent harness with LangGraph Deep Agents"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Tim Kellogg", email = "timothy.kellogg@gmail.com" }]
maintainers = [{ name = "Tim Kellogg", email = "timothy.kellogg@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "apscheduler>=3.11.2",
    "deepagents>=0.4.1",
    "discord-py>=2.6.4",
    "matplotlib>=3.8",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[project.urls]
Homepage = "https://github.com/tkellogg/open-strix"

[project.scripts]
open-strix = "open_strix.cli:main"
release = "open_strix.release:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "open_strix/builtin_skills/**/*.md",
    "open_strix/builtin_skills/**/*.py",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
